Refactor(observability): Split errors and logging
errors.py (488 lines) and structured_logging.py (446 lines) both sat
past the 400-line reviewability limit the aislop code-quality gate
applies, and each held several separable concerns.

Turn both into packages at the same dotted paths: errors splits into
taxonomy (the enums and their lookup tables), models (ErrorContext,
ClassifiedError and the classify_exception factory) and tracker (the
stateful aggregation); structured_logging splits into context (the
phase/level/record primitives), aggregation (LogAggregator) and logger
(StructuredLogger and its module-level helpers). Each package __init__
re-exports the previous names, so all import sites are unchanged.
